--HG--
branch : eportfolio

+++ modules/course_info/refresh_course.php
@@ -32,25 +32,15 @@ require_once '../../include/baseTheme.php';
 require_once 'modules/work/work_functions.php';
 require_once 'include/lib/fileManageLib.inc.php';
 
-load_js('bootstrap-datepicker');
+$data['action_bar'] = action_bar(array(
+            array('title' => $langBack,
+                  'url' => "#",
+                  'class' => 'back_btn',
+                  'icon' => 'fa-reply',
+                  'level' => 'primary-label'
+                 )));
 
-$head_content .= "
-<script type='text/javascript'>
-$(function() {
-$('#reg_date').datepicker({
-        format: 'dd-mm-yyyy',
-        language: '".$language."',
-        autoclose: true
-    });
-});
-</script>";
-
-$from_user = FALSE;
-if (isset($_GET['from_user'])) { // checks if we are coming from /modules/user/index.php
-    $from_user = $_GET['from_user'];
-}
-
-if (!$from_user) {
+if (isset($_GET['from_user'])) {
     $toolName = $langCourseInfo;
     $pageName = $langRefreshCourse;
     $navigation[] = array('url' => "index.php?course=$course_code", 'name' => $langCourseInfo);
@@ -68,7 +58,7 @@ if (isset($_POST['reg_flag'])) {
 if (isset($_POST['submit'])) {
     if (!isset($_POST['token']) || !validate_csrf_token($_POST['token'])) csrf_token_error();
     checkSecondFactorChallenge();
-    if (!$from_user) {
+    if (!isset($_GET['from_user'])) {
         $url = "$_SERVER[SCRIPT_NAME]?course=$course_code";
     } else {
         $url = "../user/index.php?course=$course_code";
@@ -93,15 +83,14 @@ if (isset($_POST['submit'])) {
     }
     if (isset($_POST['delannounces'])) {
         $output[] = delete_announcements();
-    }
-    
+    }    
     if (isset($_POST['delagenda'])) {
         $output[] = delete_agenda();
     }
     if (isset($_POST['hideworks'])) {
         $output[] = hide_work();
     }
-    if (isset($_POST['delworkssubs'])) {
+    if (isset($_POST['delworkssubs'])) {        
 	$output[] = del_work_subs();
     }            
     if (isset($_POST['purgeexercises'])) {
@@ -110,89 +99,29 @@ if (isset($_POST['submit'])) {
     if (isset($_POST['clearstats'])) {
         $output[] = clear_stats();
     }
-
-    if (($count_events = count($output)) > 0) {
-        $tool_content .= "<div class='alert alert-success'>";
-        if (!$from_user) {
-            $tool_content .= $langRefreshSuccess;
-        }
-        $tool_content .= "<ul class='listBullet'>";
-        for ($i = 0; $i < $count_events; $i++) {
-            $tool_content .= "<li>$output[$i]</li>";
-        }
-        $tool_content .= "</ul></div>";
-    }    
-} else {
-    
-    $tool_content .= action_bar(array(
-            array('title' => $langBack,
-                  'url' => "#",
-                  'class' => 'back_btn',
-                  'icon' => 'fa-reply',
-                  'level' => 'primary-label'
-                 )));
-    
-    if (!$from_user) {
-        $tool_content .= "<div class='alert alert-info'>$langRefreshInfo $langRefreshInfo_A</div>";
+    if (isset($_POST['delblogposts'])) {
+        $output[] = del_blog_posts();
     }
-    $tool_content .= "<div class='form-wrapper'>";
-            if (!$from_user) {
-                $tool_content .= "<form class='form-horizontal' role='form' action='$_SERVER[SCRIPT_NAME]?course=$course_code' method='post'>";
-            } else {
-                $tool_content .= "<form class='form-horizontal' role='form' action='$_SERVER[SCRIPT_NAME]?course=$course_code&amp;from_user=true' method='post'>";
-            }
-             $tool_content .= "<fieldset>
-            <div class='form-group'>
-                <label for='delusers' class='col-sm-2 control-label'>$langUsers</label>
-                <div class='col-sm-4 checkbox'>
-                    <label><input type='checkbox' name='delusers'>$langUserDelCourse:</label>
-                </div>
-                <div class='col-sm-3'>
-                    " . selection(array('before' => $langBefore, 'after' => $langAfter), 'reg_flag', $reg_flag, 'class="form-control"') . "
-                </div>
-                <div class='col-sm-3'>
-                    <input type='text' name='reg_date' id='reg_date' value='" .date("d-m-Y", time()) ."'>
-                </div>                
-            </div>";
-    if (!$from_user) {
-            $tool_content .= "<div class='form-group'>
-                <label for='delannounces' class='col-sm-2 control-label'>$langAnnouncements</label>
-                <div class='col-sm-10 checkbox'><label><input type='checkbox' name='delannounces'>$langAnnouncesDel</label></div>
-            </div>
-            <div class='form-group'>
-              <label for-'delagenda' class='col-sm-2 control-label'>$langAgenda</label>
-              <div class='col-sm-10 checkbox'><label><input type='checkbox' name='delagenda'>$langAgendaDel</label></div>
-            </div>
-            <div class='form-group'>
-              <label for='hideworks' class='col-sm-2 control-label'>$langWorks</label>
-                <div class='col-sm-10 checkbox'>
-                    <label><input type='checkbox' name='hideworks'>$langHideWork</label>
-                  </div>
-                <div class='col-sm-offset-2 col-sm-10 checkbox'>
-                    <label><input type='checkbox' name='delworkssubs'>$langDelAllWorkSubs</label>
-                </div>
-            </div>
-            <div class='form-group'>
-              <label for='purgeexercises' class='col-sm-2 control-label'>$langExercises</label>
-              <div class='col-sm-10 checkbox'><label><input type='checkbox' name='purgeexercises'>$langPurgeExercisesResults</label></div>
-            </div>
-            <div class='form-group'>
-              <label for='clearstats' class='col-sm-2 control-label'>$langUsage</label>
-              <div class='col-sm-10 checkbox'><label><input type='checkbox' name='clearstats'>$langClearStats</label></div>
-            </div>";
-            }
-        $tool_content .= "
-            ".showSecondFactorChallenge()."
-            <div class='col-sm-offset-2 col-sm-10'>
-            <input class='btn btn-primary' type='submit' value='$langSubmitActions' name='submit'>
-            </div>
-            </fieldset>
-            ". generate_csrf_token_form_field() ."                              
-            </form>
-            </div>";    
+    if (isset($_POST['delwallposts'])) {
+        $output[] = del_wall_posts();
+    }
+    // output results
+    if (($count_events = count($output)) > 0) {
+        $data['count_events'] = $count_events;
+        $data['output'] = $output;
+        $data['menuTypeID'] = 2;
+        view('modules.course_info.refresh_course_results', $data);
+    }    
+} else { // display form                 
+    $data['selection_date'] = selection(array('before' => $langBefore, 'after' => $langAfter), 'reg_flag', $reg_flag, 'class="form-control"');
+    $data['date_format'] = date("d-m-Y", time());
+        
+    $data['form_url'] = "$_SERVER[SCRIPT_NAME]?course_code=$course_code";
+    $data['form_url_from_user'] = "$_SERVER[SCRIPT_NAME]?course=$course_code&amp;from_user=true";
+    
+    $data['menuTypeID'] = 2;
+    view('modules.course_info.refresh_course', $data);
 }
-
-draw($tool_content, 2, null, $head_content);
 
 /**
  * 
@@ -222,7 +151,7 @@ function delete_users($date = '', $duration = '') {
                          WHERE group_id IN (SELECT id FROM `group` WHERE course_id = ?d) AND
                                user_id NOT IN (SELECT user_id FROM course_user WHERE course_id = ?d)", $course_id, $course_id);
     
-    return "<p>$langUsersDeleted</p>";
+    return "$langUsersDeleted";
 }
 
 /**
@@ -235,7 +164,7 @@ function delete_announcements() {
     global $course_id, $langAnnDeleted;
 
     Database::get()->query("DELETE FROM announcement WHERE course_id = ?d", $course_id);
-    return "<p>$langAnnDeleted</p>";
+    return "$langAnnDeleted";
 }
 
 /**
@@ -248,7 +177,7 @@ function delete_agenda() {
     global $langAgendaDeleted, $course_id;
 
     Database::get()->query("DELETE FROM agenda WHERE course_id = ?d", $course_id);
-    return "<p>$langAgendaDeleted</p>";
+    return "$langAgendaDeleted";
 }
 
 /**
@@ -261,7 +190,7 @@ function hide_doc() {
     global $langDocsDeleted, $course_id;
 
     Database::get()->query("UPDATE document SET visible=0, public=0 WHERE course_id = ?d", $course_id);
-    return "<p>$langDocsDeleted</p>";
+    return "$langDocsDeleted";
 }
 
 /**
@@ -274,7 +203,7 @@ function hide_work() {
     global $langWorksDeleted, $course_id;
 
     Database::get()->query("UPDATE assignment SET active=0 WHERE course_id = ?d", $course_id);
-    return "<p>$langWorksDeleted</p>";
+    return "$langWorksDeleted";
 }
 /**
  * 
@@ -285,19 +214,26 @@ function hide_work() {
  * @return type
  */
 function del_work_subs()  {
-	global $langAllAssignmentSubsDeleted, $webDir, $course_id, $course_code;
-        
-        $workPath = $webDir."/courses/".$course_code."/work";
+    global $langAllAssignmentSubsDeleted, $webDir, $course_id, $course_code;
 
-        $result = Database::get()->queryArray("SELECT id FROM assignment WHERE course_id = ?d", $course_id);
-        
-        foreach ($result as $row) {  
-            $secret = work_secret($row->id);
+    $workPath = $webDir."/courses/".$course_code."/work";
+
+    $result = Database::get()->queryArray("SELECT id FROM assignment WHERE course_id = ?d", $course_id);
+
+    foreach ($result as $row) {
+        $secret =  Database::get()->querySingle("SELECT secret_directory FROM assignment 
+                            WHERE course_id = ?d AND id = ?d", $course_id, $row->id);
+        if ($secret) {
+            if (is_dir("$workPath/$secret->secret_directory")) { // if exists secret directory
+                if (count(scandir("$workPath/$secret->secret_directory")) > 2) { // and is not empty
+                    move_dir("$workPath/$secret->secret_directory",
+                       "$webDir/courses/garbage/${course_code}_work_".$row->id."_$secret->secret_directory");
+                }
+            }
             Database::get()->query("DELETE FROM assignment_submit WHERE assignment_id = ?d", $row->id);
-            move_dir("$workPath/$secret",
-            "$webDir/courses/garbage/${course_code}_work_".$row->id."_$secret");
         }
-	return "<p>$langAllAssignmentSubsDeleted</p>";
+    }
+    return "$langAllAssignmentSubsDeleted";
 }
 
 /**
@@ -314,7 +250,7 @@ function purge_exercises() {
     Database::get()->query("DELETE FROM exercise_user_record WHERE eid IN
                                 (SELECT id FROM exercise WHERE course_id = ?d)",$course_id);
 
-    return "<p>$langPurgeExercisesResults</p>";
+    return "$langPurgeExercisesResults";
 }
 
 /**
@@ -329,5 +265,47 @@ function clear_stats() {
     $action = new action();
     $action->summarizeAll();
 
-    return "<div class='alert alert-info'>$langStatsCleared</div>";
+    return "$langStatsCleared";
 }
+
+/**
+ *
+ * @global type $langAllBlogPostsDeleted
+ * @return type
+ */
+function del_blog_posts() {
+    global $langBlogPostsDeleted, $course_id;
+
+    Database::get()->query("DELETE `comments` FROM `comments` INNER JOIN `blog_post` ON `comments`.`rid` = `blog_post`.`id`
+                            WHERE `comments`.`rtype` = ?s AND `blog_post`.`course_id` = ?d", 'blogpost', $course_id);
+    Database::get()->query("DELETE `rating` FROM `rating` INNER JOIN `blog_post` ON `rating`.`rid` = `blog_post`.`id`
+                            WHERE `rating`.`rtype` = ?s AND `blog_post`.`course_id` = ?d", 'blogpost', $course_id);
+    Database::get()->query("DELETE `rating_cache` FROM `rating_cache` INNER JOIN `blog_post` ON `rating_cache`.`rid` = `blog_post`.`id`
+                            WHERE `rating_cache`.`rtype` = ?s AND `blog_post`.`course_id` = ?d", 'blogpost', $course_id);
+    Database::get()->query("DELETE FROM `blog_post` WHERE `course_id` = ?d", $course_id);
+    
+    return "$langBlogPostsDeleted";
+}
+
+/**
+ *
+ * @global type $langWallPostsDeleted
+ * @return type 
+ */
+function del_wall_posts() {
+    global $langWallPostsDeleted, $course_id;
+
+    Database::get()->query("DELETE `rating` FROM `rating` INNER JOIN `wall_post` ON `rating`.`rid` = `wall_post`.`id`
+                            WHERE `rating`.`rtype` = ?s AND `wall_post`.`course_id` = ?d", 'wallpost', $course_id);
+    Database::get()->query("DELETE `rating_cache` FROM `rating_cache` INNER JOIN `wall_post` ON `rating_cache`.`rid` = `wall_post`.`id`
+                            WHERE `rating_cache`.`rtype` = ?s AND `wall_post`.`course_id` = ?d", 'wallpost', $course_id);
+    Database::get()->query("DELETE `comments` FROM `comments` INNER JOIN `wall_post` ON `comments`.`rid` = `wall_post`.`id`
+                            WHERE `comments`.`rtype` = ?s AND `wall_post`.`course_id` = ?d", 'wallpost', $course_id);
+    Database::get()->query("DELETE `wall_post_resources` FROM `wall_post_resources` INNER JOIN `wall_post` ON `wall_post_resources`.`post_id` = `wall_post`.`id`
+                            WHERE `wall_post`.`course_id` = ?d", $course_id);
+    Database::get()->query("DELETE FROM abuse_report WHERE rtype = ?s AND course_id = ?d", 'wallpost', $course_id);
+    Database::get()->query("DELETE FROM `wall_post` WHERE `course_id` = ?d", $course_id);
+
+    return "$langWallPostsDeleted";
+}
+
